+++ lib/utils/number_button.dart
@@ -4,29 +4,33 @@ typedef OnNumberButtonPressed = void Function(int number);
 
 class NumberButton extends StatelessWidget {
   const NumberButton({
+    super.key,
     required this.number,
     required this.size,
     required this.color,
     required this.onNumberTap,
     this.buttonElevation,
-    super.key,
+    this.foregroundColor,
+    this.buttonRadius,
   });
   final int number;
-  final double size;
+  final Size size;
   final Color color;
   final OnNumberButtonPressed onNumberTap;
   final double? buttonElevation;
+  final double? buttonRadius;
+  final Color? foregroundColor;
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      width: size,
-      height: size,
+      width: size.width,
+      height: size.height,
       child: ElevatedButton(
         style: ElevatedButton.styleFrom(
           backgroundColor: color,
           shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(size / 2),
+            borderRadius: BorderRadius.circular(buttonRadius ?? 100),
           ),
           elevation: buttonElevation ?? 4,
           surfaceTintColor: Colors.white,
@@ -34,13 +38,15 @@ class NumberButton extends StatelessWidget {
         onPressed: () {
           onNumberTap(number);
         },
-        child: Center(
-          child: Text(
-            number.toString(),
-            style: const TextStyle(
-              fontWeight: FontWeight.bold,
-              color: Colors.black,
-              fontSize: 30,
+        child: FittedBox(
+          child: Center(
+            child: Text(
+              number.toString(),
+              style: TextStyle(
+                fontWeight: FontWeight.bold,
+                color: foregroundColor ?? Colors.black,
+                fontSize: 30,
+              ),
             ),
           ),
         ),