+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 1.0.0
+
+### Added
+
+- Added `buttonBackgroundColor` property to customize the background color of number buttons.
+- Added `buttonForegroundColor` property to customize the foreground color of number buttons.
+- Added `buttonBorderRadius` property to specify the border radius of number buttons.
+- Added `buttonSize` property to control the size of number buttons.
+- Added `numPadVerticalSpacing` property to adjust vertical spacing between number buttons in the number pad.
+- Added `numPadHorizontalSpacing` property to adjust horizontal spacing between number buttons in the number pad.
+
 ## 0.0.6
 
-* Fix issues and remove unused appbar
-## 0.0.1
-
-* TODO: Describe initial release.
+- Fix issues and remove unused appbar